% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_cases}
\alias{calculate_cases}
\title{Calculate the number of primary, secondary, tertiary and quaternary
 dengue cases.}
\usage{
calculate_cases(incidences, sym_to_asym_ratios, n_j, N)
}
\arguments{
\item{incidences}{A numeric vector of incidence of primary, secondary,
tertiary and quaternary dengue infections.}

\item{sym_to_asym_ratios}{A list of the proportion of symptomatic primary,
secondary, tertiary and quaternary cases (in this order).}

\item{...}{Additional arguments to pass to \code{incidence_to_numbers()}}
}
\value{
Numeric.
}
\description{
Calculate the number of primary, secondary, tertiary and quaternary
 dengue symptomatic cases in a population.
}
