% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_incidences}
\alias{calculate_incidences}
\title{Calculate the incidence of primary, secondary, tertiary and quaternary
 dengue infections.}
\usage{
calculate_incidences(FOI, u_lim, l_lim)
}
\arguments{
\item{FOI}{The force of infection value. Numeric.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}
}
\value{
A list containing four different numeric vectors with the incidence
 of primary, secondary, tertiary and quaternary infections in each age group.
}
\description{
\code{calculate_incidences} calculates the incidence of primary, secondary,
tertiary and quaternary dengue infections in a population, for a given
force of infection. \emph{Incidence} is the probability that someone has a
primary (or secondary, etc..) dengue infection in a particular age group.
}
\examples{
#Assumes human population divided in 5-year age groups and 20 age groups

age_groups <- 20
# a force of infection value
a_FOI_value <- 0.031
# the lower limits of the age groups
a <- seq(0, 95, length.out = age_groups)
# the upper limits of the age groups
b <- seq(5, 100, length.out = age_groups)
# run
calculate_incidences(a_FOI_value, a, b)

}
