% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_infections}
\alias{calculate_infections}
\title{Calculate the number of primary, secondary, tertiary and quaternary
 dengue infections.}
\usage{
calculate_infections(incidences, n_j, N)
}
\arguments{
\item{incidences}{A numeric vector of incidence of primary, secondary,
tertiary and quaternary dengue infections.}

\item{n_j}{A numeric vector of the proportions of individuals in each age
group.}

\item{N}{Population size. Numeric}
}
\value{
Numeric.
}
\description{
Calculate the number of primary, secondary, tertiary and quaternary
 dengue infections in a population.
}
\examples{
incidences <- c(0.0035, 0.002, 0.00123, 0.00569)
# simulate age structure
x <- sample(1:50, age_groups, replace = TRUE)
n_j <- x / sum(x)
# run
calculate_infections(FOI, a, b, n_j)

}
