% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_R0}
\alias{calculate_R0}
\title{Calculate the Dengue reproduction number (\eqn{R_0})}
\usage{
calculate_R0(FOI, n_j, u_lim, l_lim, phis)
}
\arguments{
\item{FOI}{The force of infection value. Numeric.}

\item{n_j}{A numeric vector of the number of individuals in each age group
in the human population.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}

\item{phis}{A numeric vector of the relative infectiousness
of primary, secondary, tertiary and quaternary dengue infections.}
}
\value{
Numeric.
}
\description{
\code{calculate_R0} calculates the dengue reproduction number using the
at-equilibrium number of primary, secondary, tertiary and quaternary infections
in a population and their relative infectiousness.
}
