% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_R0_foi_and_infectiouss_bulk}
\alias{calculate_R0_foi_and_infectiouss_bulk}
\title{Calculate R0 for multiple force of infection values and infectiousness
 combinations.}
\usage{
calculate_R0_foi_and_infectiouss_bulk(foi_data, phi_combs, age_data, id_0_field,
  foi_field, pop_field, u_lim, l_lim)
}
\arguments{
\item{foi_data}{A matrix with the force of infection (FOI) estimates in one
column. The other columns contain other attributes related to the location
of the force of infection data point (e.g. longitude, latitude, population etc.).}

\item{phi_combs}{A list with the different combinations of infectiousness
of the four dengue infection.}

\item{age_data}{A matrix with the age structure of the different force of
infections in \code{foi_data}. Each row refers to an individual force of
infection estimate. The first column is the GADM adm 0 unique identifier for
the country of the force of infection data point. The remaining columns
contain the proportions of individuals in each age group.}

\item{id_0_field}{A character string of the name of the column where the
GADM adm 0 unique identifier for the country is stored. It has to be the same
in the foi_data and the age_data datasets.}

\item{foi_field}{A character string of the force of infection column name.}

\item{pop_field}{A character string of the population size column name.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}
}
\value{
A \emph{n} x \emph{m} matrix where \emph{n} is the number of force
of infection estimates and emph{m} is the number of infectiousness combinations.
}
\description{
Wrapper for repeating calculation of the dengue reproduction number over
multiple force of infection values and combination of infectiousness.
}
