% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_R0_foi_bulk}
\alias{calculate_R0_foi_bulk}
\title{Calculate R0 for multiple force of infection values.}
\usage{
calculate_R0_foi_bulk(foi_data, age_data, id_0_field, foi_field, pop_field,
  u_lim, l_lim, phis)
}
\arguments{
\item{foi_data}{A matrix with three columns. One contains the GADM adm 0
unique identifier for the country of the force of infection data point. One
contains the force of infection estimates. One contains the values of population
size for the location to which the force of infection refers.}

\item{age_data}{A matrix with the age structure of the different force of
infections in \code{foi_data}. Each row refers to an individual force of
infection estimate. The first column is the GADM adm 0 unique identifier for
the country of the force of infection data point. The remaining columns
contain the proportions of individuals in each age group.}

\item{id_0_field}{A character string of the name of the column where the
GADM adm 0 unique identifier for the country is stored. It has to be the same
in the foi_data and the age_data datasets.}

\item{foi_field}{A character string of the force of infection column name.}

\item{pop_field}{A character string of the population size column name.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}

\item{phis}{A numeric vector of the relative infectiousness
of each of the four dengue infections.}
}
\value{
A numeric vector of length equals \code{nrow(foi_data)}.
}
\description{
Wrapper for repeating calculation of the dengue reproduction number over
multiple force of infection values.
}
