% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_R0_infectiouss_bulk}
\alias{calculate_R0_infectiouss_bulk}
\title{Calculate R0 for multiple infectiousness combinations.}
\usage{
calculate_R0_infectiouss_bulk(phi_combs, ...)
}
\arguments{
\item{phi_combs}{A list with the different combinations of infectiousness
of the four dengue infection.}

\item{...}{Additional arguments to pass to \code{calculate_R0()}.}
}
\value{
A numeric vector.
}
\description{
Wrapper for repeating calculation of the dengue reproduction number over
multiple infectiousness combinations.
}
