% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGraph.R
\name{dGraph}
\alias{dGraph}
\title{Graphical Representation}
\usage{
dGraph(seq, dim = 2)
}
\arguments{
\item{seq}{sequence given as a single string, e.g. "AGTGGG" or as a FASTA type}

\item{dim}{- number of dimensions: 2 (default) or 3}

\item{genbank}{FALSE (default) - raw sequence, TRUE - accession number from GenBank}

\item{vec}{- list of unit vectors associated with nucleobases}
}
\value{
A list containing two matrices: 'coordinates' and 'graph'
}
\description{
Create graphical representation
}
