% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foi}
\alias{foi}
\title{Geolocated force of infection estimates of dengue}
\format{A data frame with 382 rows and 11 variables:
\describe{
   \item{data_id}{unique identifier}
   \item{type}{type of study where the force of infection estimate was sourced from (serology, caseReport, pseudoAbsence)}
   \item{date}{date of the study}
   \item{longitude}{longitude of the location of the study}
   \item{latitude}{latitude of the location of the study}
   \item{country}{country where the study was carried out}
   \item{ISO}{3-digits country code}
   \item{ID_0}{GADM administrative unit level 0 code}
   \item{ID_1}{GADM administrative unit level 1 code}
   \item{FOI}{force of infection estimate (0.00000000--0.06159744)}
   \item{population}{size of the human population in level 1 adm unit}
 }}
\usage{
foi
}
\description{
A dataset containing the values and other attributes of > 580
force of infection estimates of dengue. The variables are as follows:
}
\references{
\url{https://mrcdata.dide.ic.ac.uk/_dengue/dengue.php}
}
\keyword{datasets}
