% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_burden.R
\name{get_age_band_bounds}
\alias{get_age_band_bounds}
\title{\code{get_age_band_bounds} returns the lower and upper limits of an age group
character string of the format e.g. \code{band_X_Y}, where X and Y represent
the age limits.}
\usage{
get_age_band_bounds(tags)
}
\arguments{
\item{tags}{Character string of the age group}
}
\value{
Numeric.
}
\description{
\code{get_age_band_bounds} returns the lower and upper limits of an age group
character string of the format e.g. \code{band_X_Y}, where X and Y represent
the age limits.
}
