% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2DGraph.R
\name{plot2DGraph}
\alias{plot2DGraph}
\title{Plot 2D scatter plot}
\usage{
plot2DGraph(seq, dim = 2, genbank = FALSE, xlab = 'X', ylab = 'Y', main = '', colorset = c('#da186f', '#681f1c', '#ffa600', '#bc5090', '#003f5c'), show.legend = TRUE, legend.pos = 'topleft')
}
\arguments{
\item{genbank}{FALSE for regular strings, TRUE for GenBank accession numbers}

\item{colorset}{vector of colors}

\item{show.legend}{TRUE of FALSE}

\item{legend.pos}{position of the legend ('topleft' as default)}

\item{seq}{DNA sequence given as a string or as GenBank accession number}

\item{dim}{number of dimensions (2 or 3)}
}
\description{
Plot 2D scatter plot
}
