% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_burden.R
\name{calculate_hosp_cases}
\alias{calculate_hosp_cases}
\title{Calculate the number of dengue hospitalized cases.}
\usage{
calculate_hosp_cases(FOI, n_j, u_lim, l_lim, parms, weights_vec, N)
}
\arguments{
\item{FOI}{The force of infection value. Numeric.}

\item{n_j}{A numeric vector of the proportion of individuals in each age group
in the human population.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}

\item{parms}{A numeric vector of length = 4 of the proportions of primary,
secondary, tertiary and quaternary symptomatic infections requiring
hospitalization.}

\item{weights_vec}{A numeric vector of length = 4 of the proportions of
primary, secondary, tertiary and quaternary infections which are symptomatic.}

\item{N}{Size of human population. NUmeric.}
}
\value{
Numeric.
}
\description{
\code{calculate_hosp_cases} calculates the number of annual dengue symptomatic
infections (\emph{cases}) requiring hospitalization in a human population,
given a force of infection.
}
