% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_burden.R
\name{calculate_infections}
\alias{calculate_infections}
\title{Calculate the number of dengue infections.}
\usage{
calculate_infections(FOI, n_j, u_lim, l_lim)
}
\arguments{
\item{FOI}{The force of infection value. Numeric.}

\item{n_j}{A numeric vector of the number of individuals in each age group
in the human population.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}
}
\value{
Numeric.
}
\description{
\code{calculate_infections} calculates the number of annual dengue infections
in a human population, given a force of infection.
}
