% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_infectiousness_sym_2x_asym}
\alias{calculate_infectiousness_sym_2x_asym}
\title{Calculate infectiousness when symptomatic infections are twice as
 infectious as asymptomatic ones.}
\usage{
calculate_infectiousness_sym_2x_asym(prop_sym_parms)
}
\arguments{
\item{prop_sym_parms}{numeric vector of proportions of primary, secondary and
tertiary/quaternary infections which are symptomatic.}
}
\value{
Numeric.
}
\description{
\code{calculate_infectiousness_sym_2x_asym} calculates the infectiousness of
primary, secondary, tertiary and quaternary assuming that symptomatic infections
are twice as infecious as asymptomatic ones.
Infectiousness of secondary infectious is 1.
}
