% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_R0}
\alias{calculate_R0}
\title{Calculate the Dengue reproduction number (R0).}
\usage{
calculate_R0(FOI, N, infections, phis)
}
\arguments{
\item{FOI}{The force of infection value. Numeric.}

\item{infections}{A numeric vector of the number of primary, secondary,
tertiary and quaternary infections}

\item{phis}{A numeric vector of the relative infectiousness
of each dengue infection}
}
\value{
Numeric.
}
\description{
Calculate the dengue reproduction number using the at-equilibrium number of
primary, secondary, tertiary and quaternary infections in a population and
their relative infectiousness.
}
\examples{
# a force of infection value
a_FOI <- 0.031
# a vector of infection numbers
vec_infections <- c(10, 3000, 1, 42)
a vector of infectiousness weights
vec_phis <- c(1, 1, 1, 1)
# run
calculate_R0(a_Foi, vec_infections, vec_phis)

}
