% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{calculate_R0}
\alias{calculate_R0}
\title{Calculate the Dengue reproduction number (\eqn{R_0})}
\usage{
calculate_R0(FOI, N, n_j, u_lim, l_lim, phis)
}
\arguments{
\item{FOI}{The force of infection value. Numeric.}

\item{N}{Population size. Numeric.}

\item{n_j}{A numeric vector of the proportions of individuals in each age
group.}

\item{u_lim}{A numeric vector of the upper age limits of the population age
groups.}

\item{l_lim}{A numeric vector of the lower age limits of the population age
groups.}

\item{phis}{A numeric vector of the relative infectiousness
of each of the four dengue infections.}
}
\value{
Numeric.
}
\description{
Calculate the dengue reproduction number using the at-equilibrium number of
primary, secondary, tertiary and quaternary infections in a population and
their relative infectiousness.
}
