% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{incidences_to_numbers}
\alias{incidences_to_numbers}
\title{Convert incidences into numbers.}
\usage{
incidences_to_numbers(incidences, N, n_j)
}
\arguments{
\item{incidences}{A numeric vector of the incidence of primary, secondary,
tertiary and quaternary dengue infections.}

\item{N}{Population size. Numeric.}

\item{n_j}{A numeric vector of the proportions of individuals in each age
group.}
}
\value{
Numeric.
}
\description{
\code{incidences_to_numbers} calculates the \emph{number} of people
(infected or displaying disease symptoms) from \emph{incidence} estimates.
}
