% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating_R0.R
\name{incidences_to_numbers}
\alias{incidences_to_numbers}
\title{Convert incidences into numbers.}
\usage{
incidences_to_numbers(incidences, n_j)
}
\arguments{
\item{incidences}{A numeric vector of the incidence rates of primary, secondary,
tertiary or quaternary dengue infections in each age group.}

\item{n_j}{A numeric vector of the proportion of individuals in each age group
in the human population.}
}
\value{
Numeric.
}
\description{
\code{incidences_to_numbers} calculates the \emph{numbers} of people from different
age groups experiencing their primary, secondary, tertiary or quaternary
dengue infection, from estimates of the incidence \emph{rates} of primary, secondary,
tertiary or quaternary infections in each age group.
}
